requirements.txt
astroid==2.4.2
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
isort==5.5.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.21
protobuf==3.13.0
pycodestyle==2.6.0
pylint==2.6.0
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
